requirements.txt
pika
python-dotenv
boto3
PyPDF2
google-generativeai
